+++ bin/generated-test-sources/annotations/com/vajrapulse/vortex/benchmark/jmh_generated/LatencyBenchmark_jmhType.java
@@ -1,0 +1,4 @@
+package com.vajrapulse.vortex.benchmark.jmh_generated;
+public class LatencyBenchmark_jmhType extends LatencyBenchmark_jmhType_B3 {
+}
+