+++ bin/generated-test-sources/annotations/com/vajrapulse/vortex/benchmark/jmh_generated/MicroBatcherBenchmark_jmhType.java
@@ -1,0 +1,4 @@
+package com.vajrapulse.vortex.benchmark.jmh_generated;
+public class MicroBatcherBenchmark_jmhType extends MicroBatcherBenchmark_jmhType_B3 {
+}
+