+++ bin/generated-test-sources/annotations/com/vajrapulse/vortex/jmh_generated/SubmitSyncBenchmark_jmhType.java
@@ -1,0 +1,4 @@
+package com.vajrapulse.vortex.jmh_generated;
+public class SubmitSyncBenchmark_jmhType extends SubmitSyncBenchmark_jmhType_B3 {
+}
+